// frontend/src/Types/CompanyCashFlow.d.ts
export interface CompanyCashFlow {
    "date": string;
    "symbol": string;
    "reportedCurrency": string;
    "cik": string;
    "filingDate": string;
    "acceptedDate": string;
    "fiscalYear": string;
    "period": string;
    "netIncome": number;
    "depreciationAndAmortization": number;
    "deferredIncomeTax": number;
    "stockBasedCompensation": number;
    "changeInWorkingCapital": number;
    "accountsReceivables": number;
    "inventory": number;
    "accountsPayables": number;
    "otherWorkingCapital": number;
    "otherNonCashItems": number;
    "netCashProvidedByOperatingActivities": number;
    "investmentsInPropertyPlantAndEquipment": number;
    "acquisitionsNet": number;
    "purchasesOfInvestments": number;
    "salesMaturitiesOfInvestments": number;
    "otherInvestingActivities": number;
    "netCashProvidedByInvestingActivities": number;
    "netDebtIssuance": number;
    "longTermNetDebtIssuance": number;
    "shortTermNetDebtIssuance": number;
    "netStockIssuance": number;
    "netCommonStockIssuance": number;
    "commonStockIssuance": number;
    "commonStockRepurchased": number;
    "netPreferredStockIssuance": number;
    "netDividendsPaid": number;
    "commonDividendsPaid": number;
    "preferredDividendsPaid": number;
    "otherFinancingActivities": number;
    "netCashProvidedByFinancingActivities": number;
    "effectOfForexChangesOnCash": number;
    "netChangeInCash": number;
    "cashAtEndOfPeriod": number;
    "cashAtBeginningOfPeriod": number;
    "operatingCashFlow": number;
    "capitalExpenditure": number;
    "freeCashFlow": number;
    "incomeTaxesPaid": number;
    "interestPaid": number;
}